api_yamdb/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt
django-filter==22.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref==3.3.2
pytz==2020.1
sqlparse==0.3.1
